stand_alone/basic_kafka_example/requirements.txt
certifi>=2021.10.8
kafka-python>=2.0.2
